NEWS-29 Poprawienie komentarzy z PR
