+++ lib/l10n/intl_en.arb
@@ -6,7 +6,6 @@
   "darkMode": "Dark mode",
   "newest": "Newest",
   "next": "Next",
-  "news": "News",
   "profile": "Profile",
   "mostViewed": "Most Viewed",
   "mostShared": "Most Shared",
