+++ lib/l10n/intl_pl.arb
@@ -6,7 +6,6 @@
   "darkMode": "Tryb ciemny",
   "newest": "Najnowsze",
   "next": "Dalej",
-  "news": "News",
   "profile": "Profil",
   "mostViewed": "Najczęściej oglądane",
   "mostShared": "Najczęściej udostępniane",
