+++ lib/src/ui/pages/dashboard/news/secondary_news/secondary_news_list_item.dart
@@ -1,6 +1,6 @@
 import 'package:allthenews/src/ui/common/util/dimens.dart';
 import 'package:allthenews/src/ui/common/widget/dot_separator.dart';
-import 'package:allthenews/src/ui/pages/feed/news/secondary_news/secondary_news_list_entity.dart';
+import 'package:allthenews/src/ui/pages/dashboard/news/secondary_news/secondary_news_list_entity.dart';
 import 'package:allthenews/src/ui/pages/web_view/web_view_page.dart';
 import 'package:flutter/material.dart';
